Added automatic checkpoint to nemo file method
